Arreglos EP (Módulos Integrados)

+++ KB_Guadalupana/Views/Sesion/ReportesAdmin/ReporteAdmin3.aspx.cs
@@ -21,9 +21,9 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
 
         string sesion;
 
-        int caja1, cuentasQ1, CooperativasQ1, CP1, IN1, Inmuebles1, vehiculo1, maquinaria1, computo1, TotalQ1, salas1, comedor1, tele, Es1, L1, S1, Est1, Ref1, Mov1, otr1;
-        int cuentasD1, CooperativasD1, TotalD1;
-        int pp1, TotalQ2, patrimonio1, pres1, tc1, OD1, pas1, pass2, pass3;
+        double caja1, cuentasQ1, CooperativasQ1, CP1, IN1, Inmuebles1, vehiculo1, maquinaria1, computo1, TotalQ1, salas1, comedor1, tele, Es1, L1, S1, Est1, Ref1, Mov1, otr1;
+        double cuentasD1, CooperativasD1, TotalD1;
+        double pp1, TotalQ2, patrimonio1, pres1, tc1, OD1, pas1, pass2, pass3;
         string completo, nombre1, nombre2, apellido1, apellido2;
 
 
@@ -118,7 +118,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCajas(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                caja1 = Convert.ToInt32(var1[0]);
+                caja1 = Convert.ToDouble(var1[0]);
                 EfectivoCaja.Value = caja1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -136,7 +136,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoQ(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                cuentasQ1 = Convert.ToInt32(var1[0]);
+                cuentasQ1 = Convert.ToDouble(var1[0]);
                 CuentasBancos.Value = cuentasQ1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -148,7 +148,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoD(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                cuentasD1 = Convert.ToInt32(var1[0]);
+                cuentasD1 = Convert.ToDouble(var1[0]);
                 CuentasBancos1.Value = cuentasD1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -160,7 +160,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCD(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                CooperativasD1 = Convert.ToInt32(var1[0]);
+                CooperativasD1 = Convert.ToDouble(var1[0]);
                 CuentasCope1.Value = CooperativasD1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -172,7 +172,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCQ(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                CooperativasQ1 = Convert.ToInt32(var1[0]);
+                CooperativasQ1 = Convert.ToDouble(var1[0]);
                 CuentasCope.Value = CooperativasQ1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -184,7 +184,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCP(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                CP1 = Convert.ToInt32(var1[0]);
+                CP1 = Convert.ToDouble(var1[0]);
                 CuentasPC.Value = CP1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -196,7 +196,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoIN(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                IN1 = Convert.ToInt32(var1[0]);
+                IN1 = Convert.ToDouble(var1[0]);
                 Inventario.Value = IN1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -208,7 +208,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoInmueble(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                Inmuebles1 = Convert.ToInt32(var1[0]);
+                Inmuebles1 = Convert.ToDouble(var1[0]);
                 Inmueble.Value = Inmuebles1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -220,7 +220,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoVehiculo(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                vehiculo1 = Convert.ToInt32(var1[0]);
+                vehiculo1 = Convert.ToDouble(var1[0]);
                 Vehiculo.Value = vehiculo1.ToString("N1", CultureInfo.CurrentCulture); ;
             }
         }
@@ -232,7 +232,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoMaquinaria(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                maquinaria1 = Convert.ToInt32(var1[0]);
+                maquinaria1 = Convert.ToDouble(var1[0]);
                 Maquinaria.Value = maquinaria1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -244,7 +244,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoComputo(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                computo1 = Convert.ToInt32(var1[0]);
+                computo1 = Convert.ToDouble(var1[0]);
                 Computo.Value = computo1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -256,7 +256,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoSala(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                salas1 = Convert.ToInt32(var1[0]);
+                salas1 = Convert.ToDouble(var1[0]);
                 sala.Value = salas1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -268,7 +268,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoComedor(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                comedor1 = Convert.ToInt32(var1[0]);
+                comedor1 = Convert.ToDouble(var1[0]);
                 comedor.Value = comedor1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -280,7 +280,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoTV(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                tele = Convert.ToInt32(var1[0]);
+                tele = Convert.ToDouble(var1[0]);
                 TV.Value = tele.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -292,7 +292,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoES(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                Es1 = Convert.ToInt32(var1[0]);
+                Es1 = Convert.ToDouble(var1[0]);
                 EquipoS.Value = Es1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -304,7 +304,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoL(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                L1 = Convert.ToInt32(var1[0]);
+                L1 = Convert.ToDouble(var1[0]);
                 lava.Value = L1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -316,7 +316,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoSec(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                S1 = Convert.ToInt32(var1[0]);
+                S1 = Convert.ToDouble(var1[0]);
                 SECA.Value = S1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -328,7 +328,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoEst(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                Est1 = Convert.ToInt32(var1[0]);
+                Est1 = Convert.ToDouble(var1[0]);
                 Estufa.Value = Est1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -340,7 +340,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoRefri(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                Ref1 = Convert.ToInt32(var1[0]);
+                Ref1 = Convert.ToDouble(var1[0]);
                 Refri.Value = Ref1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -352,7 +352,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoTel(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                Mov1 = Convert.ToInt32(var1[0]);
+                Mov1 = Convert.ToDouble(var1[0]);
                 Movil.Value = Mov1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -364,7 +364,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoOtros(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                otr1 = Convert.ToInt32(var1[0]);
+                otr1 = Convert.ToDouble(var1[0]);
                 otross.Value = otr1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -383,7 +383,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCP1(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                pp1 = Convert.ToInt32(var1[0]);
+                pp1 = Convert.ToDouble(var1[0]);
                 CuentaPP.Value = pp1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -395,7 +395,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCuentaspp(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                pres1 = Convert.ToInt32(var1[0]);
+                pres1 = Convert.ToDouble(var1[0]);
                 Pres.Value = pres1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -407,7 +407,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCuentaTarjeta(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                tc1 = Convert.ToInt32(var1[0]);
+                tc1 = Convert.ToDouble(var1[0]);
                 TC.Value = tc1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -419,7 +419,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCuentaOD(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                OD1 = Convert.ToInt32(var1[0]);
+                OD1 = Convert.ToDouble(var1[0]);
                 ode.Value = OD1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -431,7 +431,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCuentapasc(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                pas1 = Convert.ToInt32(var1[0]);
+                pas1 = Convert.ToDouble(var1[0]);
                 pasc.Value = pas1.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -443,7 +443,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCuentaFena(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                pass2 = Convert.ToInt32(var1[0]);
+                pass2 = Convert.ToDouble(var1[0]);
                 Text4.Value = pass2.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
@@ -455,7 +455,7 @@ namespace KB_Guadalupana.Views.Sesion.ReportesAdmin
             string[] var1 = sn.consultarconcampoCuentaFenaINver(sesion);
             for (int i = 0; i < var1.Length; i++)
             {
-                pass3 = Convert.ToInt32(var1[0]);
+                pass3 = Convert.ToDouble(var1[0]);
                 Text1.Value = pass3.ToString("N1", CultureInfo.CurrentCulture);
             }
         }
